requirements.txt
flask
openai
azure-core
azure-search-documents
gunicorn